add read and send interfaces and refactor
