+++ src/main/java/ReadInterface.java
@@ -1,0 +1,6 @@
+import java.io.IOException;
+import java.net.Socket;
+
+public interface ReadInterface {
+  RequestParameters getRequest(Socket socket, String directoryPath) throws IOException;
+}
