+++ src/main/java/SendInterface.java
@@ -1,0 +1,6 @@
+import java.io.IOException;
+import java.net.Socket;
+
+public interface SendInterface {
+  void send(ResponseParameters responseParameters, Socket socket) throws IOException;
+}
