--- src/test/java/ShutdownHookTest.java
@@ -1,10 +1,0 @@
-import org.junit.jupiter.api.Test;
-
-public class ShutdownHookTest {
-
-  @Test
-  void shutdownHookPerformsSafeShutdown() {
-    assert(false);
-  }
-
-}
